requirements.txt
streamlit
scikit-learn
joblib
pillow
numpy